Update windows friendliness of resolver package
