+++ service/resolv/interface.go
@@ -1,0 +1,8 @@
+package resolv
+
+type resolv interface {
+	Clean()
+	Configure()
+	New() Resolv
+	Status() bool
+}
