+++ service/resolv/types.go
@@ -1,0 +1,8 @@
+package resolv
+
+type Resolv struct {
+	Data string `yaml:"contents"`
+	File string `yaml:"file"`
+	Folder string `yaml:"folder"`
+	Name string `yaml:"name"`
+}
